Added controller and routes layers
