--- GET_BOOKS.http
@@ -1,1 +1,0 @@
-http://localhost:8000/api/book
