+++ book-review-api/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "bare-minimum node js + type script template",
   "main": "index.js",
-  "type": "module",
+  "type": "commonjs",
   "scripts": {
     "build": "npx tsc",
     "start": "node ./dist/index.js",
@@ -18,8 +18,8 @@
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
-    "@types/node": "^17.0.23",
+    "@types/node": "^17.0.45",
     "nodemon": "^2.0.15",
-    "typescript": "^4.6.3"
+    "typescript": "^4.9.5"
   }
 }
