+++ book-review-api/readme.md
@@ -1,1 +1,13 @@
-Node Js, Type Script - Bare minimum set up
+Project Structure
+
+src/
+├── controllers/
+│ ├── index.ts
+│ ├── user.controller.ts
+│ ├── auth.controller.ts
+│ └── ...
+└── routes/
+├── index.ts
+├── user.routes.ts
+├── auth.routes.ts
+└── ...
