+++ book-review-api/rest-client/GET_BOOKS.http
@@ -1,0 +1,1 @@
+http://localhost:8000/api/books
