+++ book-review-api/src/routes/bookRoutes.ts
@@ -1,0 +1,9 @@
+import { Router } from "express";
+import { BookController } from "../controllers";
+
+const router = Router();
+const bookController = new BookController();
+
+router.get("/", bookController.getBooks);
+
+export { router as bookRoutes };
