+++ package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "typescript": "^5.0.4"
+  }
+}
